keepkey: merge plugin.py into keepkey.py
(remnants of separating trezor and keepkey)
